fix(smart-replies): fix bloat, infinite retries, and type errors
**Critical bugs fixed:**

1. **Embedding bloat** (app_database.dart:130)
   - Disabled Drift SQL logging to prevent 1536-float embeddings from flooding logs
   - Was causing massive performance degradation and log spam

2. **Type cast error** (smart_reply_service.dart:99)
   - Fixed: `type '_Map<Object?, Object?>' is not a subtype of type 'Map<String, dynamic>'`
   - Use `Map<String, dynamic>.from()` to properly convert Cloud Function response

3. **Limit validation error** (semantic_search_service.dart:77)
   - Fixed: `'limit' must be an integer between 1 and 20`
   - Explicitly cast limit to int with `.toInt()` for Python Cloud Function validation

**Impact:**
- Eliminates infinite retry loop
- Fixes smart reply generation failures
- Massive performance improvement (no embedding logging)
- Vector search now works correctly

+++ lib/features/smart_replies/domain/services/semantic_search_service.dart
@@ -74,7 +74,7 @@ class SemanticSearchService {
             .call<Map<String, dynamic>>({
           'conversationId': conversationId,
           'queryEmbedding': incomingMessage.embedding,
-          'limit': limit,
+          'limit': limit.toInt(), // Ensure int type for Python validation
         });
 
         final data = result.data;
